googlestorage/requirements.txt
requests
google-cloud-pubsub
cloudevents
